Extract a PluginHook component and use it where appropriate
