+++ __tests__/src/components/WindowTopMenu.test.js
@@ -22,10 +22,11 @@ function createWrapper(props) {
 describe('WindowTopMenu', () => {
   it('renders all needed elements', () => {
     const wrapper = createWrapper();
+    console.log(wrapper.debug());
     expect(wrapper.find(Menu).length).toBe(1);
     expect(wrapper.find(WindowThumbnailSettings).length).toBe(1);
     expect(wrapper.find(WindowViewSettings).length).toBe(1);
-    expect(wrapper.find('PluginHook').length).toBe(1);
+    expect(wrapper.find('PluginHookWithHeader').length).toBe(1);
   });
 
   it('passes windowId to <WindowThumbnailSettings/>', () => {
