+++ src/components/BackgroundPluginArea.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ns from '../config/css-ns';
+import { PluginHook } from './PluginHook';
 
 /** invisible area where background plugins can add to */
-export const BackgroundPluginArea = ({ PluginComponents }) => (
+export const BackgroundPluginArea = props => (
   <div className={ns('background-plugin-area')} style={{ display: 'none' }}>
-    { renderPlugins(PluginComponents) }
+    <PluginHook {...props} />
   </div>
 );
 
@@ -16,8 +17,3 @@ BackgroundPluginArea.propTypes = {
 BackgroundPluginArea.defaultProps = {
   PluginComponents: [],
 };
-
-/** render single plugin */
-const renderPlugin = (PluginComponent, key) => <PluginComponent key={key} />;
-/** render mulitple plugins */
-const renderPlugins = PluginComponents => PluginComponents.map(renderPlugin);
