+++ src/components/WindowTopMenu.js
@@ -4,19 +4,16 @@ import ListSubheader from '@material-ui/core/ListSubheader';
 import PropTypes from 'prop-types';
 import WindowThumbnailSettings from '../containers/WindowThumbnailSettings';
 import WindowViewSettings from '../containers/WindowViewSettings';
+import { PluginHook } from './PluginHook';
 import ns from '../config/css-ns';
 
 /** Renders plugins */
-function PluginHook(props) {
+function PluginHookWithHeader(props) {
   const { PluginComponents, t } = props; // eslint-disable-line react/prop-types
   return PluginComponents ? (
     <>
       <ListSubheader role="presentation" tabIndex="-1">{t('windowPluginButtons')}</ListSubheader>
-      {
-        PluginComponents.map((PluginComponent, index) => (
-          <PluginComponent {...props} key={index} /> // eslint-disable-line react/no-array-index-key
-        ))
-      }
+      <PluginHook {...props} />
     </>
   ) : null;
 }
@@ -57,7 +54,7 @@ export class WindowTopMenu extends Component {
         >
           <WindowViewSettings windowId={windowId} handleClose={handleClose} />
           <WindowThumbnailSettings windowId={windowId} handleClose={handleClose} />
-          <PluginHook {...this.props} />
+          <PluginHookWithHeader {...this.props} />
         </Menu>
       </>
     );
