+++ src/components/WorkspaceControlPanelButtons.js
@@ -4,16 +4,7 @@ import WorkspaceAddButton from '../containers/WorkspaceAddButton';
 import WorkspaceMenuButton from '../containers/WorkspaceMenuButton';
 import WorkspaceOptionsButton from '../containers/WorkspaceOptionsButton';
 import WindowListButton from '../containers/WindowListButton';
-
-/** Renders plugins */
-const PluginHook = (props) => {
-  const { PluginComponents } = props; // eslint-disable-line react/prop-types
-  return PluginComponents ? (
-    PluginComponents.map((PluginComponent, index) => (
-      <PluginComponent {...props} key={index} /> // eslint-disable-line react/no-array-index-key
-    ))
-  ) : null;
-};
+import { PluginHook } from './PluginHook';
 
 /**
  *
